Set default of 'filter-pristine-tar' for gbp import-orig to 'False'.
Set the default of the 'filter-pristine-tar' option to 'False' for the
'gbp import-orig' command instead of 'True'.  Unless it's being repacked
or otherwise changed, but causes an error about not being able to create
a symlink from a 'gpb' version of the archive to the original version
of the archive.  Reference also the following bug in Debian:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=558777